codegen: correct TUBITAK1 name constraints.
The content was misencoded, resulting in a permitted subtree base name
that's an otherName type general name with a nonsensical value. These
types of general name are not supported by webpki, making the name
constraint functionally a no-op.

The updated encoding correctly specifies a permitted subtree with
a dNSName type base general name with the value ".tr".