codegen: use CCADB as the source of truth.
Prior to this commit the `tests/codegen.rs` generator used
https://mkcert.org as its source of truth for trusted root metadata.
This commit replaces that source of truth (and accompanying generator
code) to use https://ccadb.org instead.

The Common CA Database (CCADB) has emerged as a multi-stakeholder
repository for information about certificate authorities participating
in the trust stores maintained by CCADB root store operators.

The `IncludedCACertificateReportPEMCSV` report made available by CCADB
is a great replacement for the needs of webpki-roots:

* it allows us to filter by roots that are trusted for TLS.
* it allows us to filter by "distrust after" dates.
* it allows us to generate imposed name constraints automatically.

This removes the need to maintain a separate distrust list in
webpki-roots, or a separate manually curated imposed name constraints
set.

To minimize the trust surface of webpki-roots we take care to pin the
trust anchor used to fetch the CCADB CSV to the trust anchor in use
today for serving  https://ccadb-public.secure.force.com/, helping
minimize the risk of person-in-the-middle attack. Note that we are not
pinning the leaf/intermediates in use, just the expected root.