webpki-root-certs: add initial crate, code generation
This commit adds a new crate, `webpki-root-certs`, to the project
workspace. Compared to `webpki-roots`, this crate contains the full
self-signed DER X.509 certificate of each trust anchor in
a `TLS_SERVER_ROOT_CERTS` slice.

This is done in a separate crate from `webpki-roots` because the X.509
representation is quite large compared to the `webpki`-specific
`TrustAnchor` type. However, in some circumstances (e.g. interfacing
with non-webpki consumers like platform verifiers) the full self-signed
certificates are required. Users with this requirement can depend on
this crate in addition to/or in place of `webpki-roots`.

Care is taken to push regular users of `rustls` towards the pre-existing
crate.