webpki-ccadb: rework TLS distrust after handling
Previously we took the aggressive stance of removing a root as soon as
it was described in CCADB as having a distrust for TLS after date.
There's a good argument[0] that this is more disruptive than helpful.

Instead, remove roots with a distrust after date only once that date
plus a grace period of 398 days has past. The grace period is equal to
the maximum allowable subscriber certificate lifetime based on the
CA/Browser Forum's baseline requirements.

Since we also want to apply some simple local policy to exclude roots
unconditionally a simple block list based on the CCADB SHA256
fingerprint is also added to override the above logic to exclude a root
when required.

One FP for the GLOBALSIGN 2020 root is added to the block list. We
removed this root early because it only has <100 valid certificates in
the wild. In this case waiting for the distrust date + grace period
would mean keeping the root until Aug 2nd 2025.

[0]: https://sslmate.com/blog/post/entrust_distrust_more_disruptive_than_intended